cw2217: use -40degress to determine if the battery is in present
the device can't bootup if main_battery is not connected
but if flip battery not connected, the device can bootup and
the flip_batt_fg also can bringup. the cw2217 fg has no way to determine whether
the battery is present.

1. For flip_batt the temp NTC is in batt-package, if get a -40degress, it means
the battery is not present.
2. remove forcing the flip_battery temp is 25degerss if battery is not present.
   because factory can check flip_batt temp and present by CQA.
   it's easy to find the problem if showing the real temp

Change-Id: I252629db18f0ad8fd6438770135498d8747b0ef4
Signed-off-by: liuyc20 <liuyc20@lenovo.com>
Reviewed-on: https://gerrit.mot.com/2851838
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Guocheng Wang <wanggc6@motorola.com>
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key