charge: workaround for nu2115a abnormal enable disvacdrv
why:
The reason is that the last bit (DIS_ACDRV) of the 0x2F register
of the NU2115 was set to 1 after the OTG was unplugged,
which resulted in both the wired charging and
wireless charging paths being turned off.

Checked the code logic, as well as in the i2c bottom function
to add the print, did not see to operate this bit
in the normal OTG plug state.

Change-Id: Ia8aec511f7c6406f1fa391d78087440ca6f52d7d
Signed-off-by: huwenbo <wingt122@motorola.com>
Reviewed-on: https://gerrit.mot.com/3028065
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Wei Xu <xuwei9@lenovo.com>
Reviewed-by: Xiangpo Zhao <zhaoxp3@motorola.com>
Submit-Approved: Jira Key